litellm/pyproject.toml
[project]
name = "nlitellm-proxy"
version = "0.1.0"
description = "A LiteLLM proxy with pre-configured models and an update tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm[proxy]>=1.74.0",
    "prisma",
    "PyYAML",
    "python-dotenv",
]

[project.scripts]
nlitellm-proxy = "main:run_proxy"
nlitellm-update = "fetch_models_litellm:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["/"]
include = [
    "main.py",
    "fetch_models_litellm.py",
    "config.yaml"
]

